# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OmniRouter backend exposing cache management, health\nchecks, audit trail, and settings \u0026 feature flag endpoints.",
        "title": "OmniRouter API",
        "termsOfService": "https://github.com/vinaycharlie01/shroute",
        "contact": {
            "name": "Vinay Kumar V",
            "url": "https://github.com/vinaycharlie01"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/cache/entries": {
            "get": {
                "description": "Returns cache entries matching the given key prefix, up to the requested limit.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cache"
                ],
                "summary": "List cache entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key prefix to filter by",
                        "name": "prefix",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum entries to return (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.cacheListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/cache/flush": {
            "post": {
                "description": "Removes cache entries matching the given prefix. Pass all=true with no prefix to flush the entire Redis database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cache"
                ],
                "summary": "Flush cache entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key prefix to flush",
                        "name": "prefix",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "When true and no prefix, flushes entire cache",
                        "name": "all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.cacheFlushResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/cache/stats": {
            "get": {
                "description": "Returns aggregate cache hit/miss counters, memory usage, and entry count.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cache"
                ],
                "summary": "Get cache statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.cacheStatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/settings": {
            "get": {
                "description": "Returns all stored application settings across every category.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "List all settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.settingsListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/settings/flags": {
            "get": {
                "description": "Returns all feature flags and their current enabled state.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "List feature flags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.flagsListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/settings/flags/{name}": {
            "get": {
                "description": "Returns a single feature flag by name, including its enabled state and seeded default.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get a feature flag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flag name (e.g. CACHE_ENABLED, PII_REDACTION_ENABLED)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.flagResponse"
                        }
                    },
                    "404": {
                        "description": "Flag not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Enables or disables a feature flag by name. PII flags are opt-in — do not enable by default.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Set a feature flag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flag name (e.g. CACHE_ENABLED)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Enabled state to apply",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.setFlagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.flagResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/settings/{key}": {
            "get": {
                "description": "Returns a single setting document by its well-known key.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get a setting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Setting key (e.g. log_level, theme)",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.settingResponse"
                        }
                    },
                    "400": {
                        "description": "Unknown setting key",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Setting not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Creates or updates the JSON value for a well-known setting key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Set a setting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Setting key (e.g. log_level, theme)",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "JSON value to store",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.setValueRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.settingResponse"
                        }
                    },
                    "400": {
                        "description": "Unknown key, empty value, or invalid JSON",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes a setting by its well-known key. Returns 204 on success.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Delete a setting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Setting key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Unknown setting key",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.cacheEntryResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "ttl": {
                    "type": "string"
                }
            }
        },
        "handlers.cacheFlushResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.cacheListResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.cacheEntryResponse"
                    }
                }
            }
        },
        "handlers.cacheStatsResponse": {
            "type": "object",
            "properties": {
                "entry_count": {
                    "type": "integer"
                },
                "hits": {
                    "type": "integer"
                },
                "misses": {
                    "type": "integer"
                },
                "size_bytes": {
                    "type": "integer"
                }
            }
        },
        "handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.flagResponse": {
            "type": "object",
            "properties": {
                "default_value": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handlers.flagsListResponse": {
            "type": "object",
            "properties": {
                "flags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.flagResponse"
                    }
                }
            }
        },
        "handlers.setFlagRequest": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                }
            }
        },
        "handlers.setValueRequest": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "object"
                }
            }
        },
        "handlers.settingResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "object"
                }
            }
        },
        "handlers.settingsListResponse": {
            "type": "object",
            "properties": {
                "settings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.settingResponse"
                    }
                }
            }
        }
    }
}